requirements.txt
appnope==0.1.3
asttokens==2.2.0
backcall==0.2.0
camelot-py==0.10.1
cffi==1.15.1
chardet==5.0.0
charset-normalizer==3.0.1
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
fpdf2==2.6.0
ghostscript==0.7
humanfriendly==10.0
img2pdf==0.4.4
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
JPype1==1.4.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
nltk==3.7
numpy==1.23.5
ocrmypdf==14.0.1
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.2
parso==0.8.3
pdf2image==1.16.0
pdfminer.six==20221105
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==6.2.4
Pillow==9.3.0
pluggy==1.0.0
prompt-toolkit==3.0.33
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.11.2
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.6
regex==2022.10.31
reportlab==3.6.12
six==1.16.0
stack-data==0.6.2
tabulate==0.9.0
tk==0.1.0
tqdm==4.64.1
traitlets==5.6.0
typing_extensions==4.4.0
wcwidth==0.2.5
